Rename diff-diveresults-results to *-table
