+++ projects/planner/src/app/diff/diveresults/diff-diveresults.component.html
@@ -8,7 +8,7 @@
 
     <div class="card-body card-minheight">
         <div class="pb-3">
-            <app-diff-diveresults-results></app-diff-diveresults-results>
+            <app-diff-diveresults-table></app-diff-diveresults-table>
         </div>
     </div>
 </div>
