+++ projects/planner/src/app/diff/diveresults/table/diff-diveresults-table.component.scss
