+++ projects/planner/src/app/diff/diveresults/table/diff-diveresults-table.component.ts
@@ -9,9 +9,9 @@ import {Logger} from '../../../shared/Logger';
 import {TextConstants} from '../../../shared/TextConstants';
 
 @Component({
-    selector: 'app-diff-diveresults-results',
-    templateUrl: './diff-diveresults-results.component.html',
-    styleUrls: ['./diff-diveresults-results.component.scss', '../../diff.component.scss']
+    selector: 'app-diff-diveresults-table',
+    templateUrl: './diff-diveresults-table.component.html',
+    styleUrls: ['./diff-diveresults-table.component.scss', '../../diff.component.scss']
 })
 export class DiveResultsResultsDifferenceComponent {
     private readonly cnsDifferenceUnderMinusOneThousand = '< -1000';
